Rename ui-plugins.ts utils file
